Reworked CollisionHandler to work with CollisionObjects instead of Objects that extends CHShapes. Removed CHShapes from Shape interface. Added CollisionObject, which is basically a shape that stores an array of any other CollisionObjects it collides with. The CollisionHandler calls .collides on every CollisionObject. It is abstract, because it doesn't have a shape on its own.
Added CollisionCircle, witch is a subclass of CollisionObject and extends Shape.Circle. 

InfectoGameObject now has a pointer to a CollisionObject, which has a shape. It is this object that is added to the CollisionHandler if needed. The owner of a CollisionObject is given in the .owner.

Removed Human.java

Discovered a bug where, when humans collide their HPbar doesn't move with the sprite. It looks funny.